# Vectorize equalize transformation