Closes https://github.com/pytorch/vision/issues/3173

This PR vectorizes the histograms computation (over batches and over channels) in the `equalize` transformation.