@NicolasHug 2021-02-01
Thanks for the review @fmassa .

Interestingly, this PR does not seem to lead to any improvement (quite the opposite actually). Locally on my laptop I get the following for a `64x3x128x128` tensor:

```
This PR: 1.51 s ± 88.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
Master:  42.1 ms ± 436 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

Do you think it's worth investigating further where the slowdown might come from, or should we just close this PR and the original issue?