@fmassa 2021-02-02
> Do you think it's worth investigating further where the slowdown might come from, or should we just close this PR and the original issue?

I think the 20x slowdown on a CPU is a bit surprising. I would have expected that it could be a bit slower on CPU and faster on the GPU.

I would say to investigate this a bit further to see where the slowdown comes from -- is it from the histogram creation? If yes, then we could probably leave the histogram creation in a for loop and then perform the batch indexing as we currently have in this PR.