@NicolasHug 2021-02-02
The slowdown does seem to come from the histogram creation, which amounts for the vast majority of the computation time. As a result I'm not sure we can expect a significant improvement even if we just use batch indexing (we'd still need to stack the histograms anyway, and the execution time is dominated by the histogram computation).

Note that on GPU the difference is smaller, but master is still faster.
Histogram creation for a `(32, 3, 64, 64)` image:
```
This PR: 14.6 ms ± 616 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
Master:  8.81 ms ± 87 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```


On top of that the solution in this PR is much more memory-consuming, so we can probably just close this PR?