@fmassa 2021-02-16
Thanks for investigating this @NicolasHug !